+++ R/boxes.R
@@ -6,8 +6,7 @@
 #' @param n Number of boxes
 #' @param perc Relationship between box size and colour. Value between 0 and 1
 #' where 0 represents randomness and 1 perfect identical. Default 0.1.
-#' @param col_palette Vector of colours. Default "DarkMint" colour palette from
-#' rcartocolor.
+#' @param col_palette Vector of colours. Default `"#D2FBD4" "#A5DBC2" "#7BBCB0" "#559C9E" "#3A7C89" "#235D72" "#123F5A"`.
 #' @param bg_col Background colour. Default "black".
 #' @param s Seed value. Default 1234.
 #' @return A ggplot object.
@@ -18,7 +17,7 @@
 
 boxes <- function(n = 100,
                   perc = 0.1,
-                  col_palette = rcartocolor::carto_pal(n = 7, "DarkMint"),
+                  col_palette = c("#D2FBD4", "#A5DBC2", "#7BBCB0", "#559C9E", "#3A7C89", "#235D72", "#123F5A"),
                   bg_col = "black",
                   s = 1234) {
   plot_data <- withr::with_seed(