+++ R/mosaic.R
@@ -3,7 +3,7 @@
 #' This function generates a generative art ggplot object from voronoi tiles.
 #'
 #' @param n Number of points to generate tiles from. Default 100.
-#' @param fill_cols Vector of colours to fill tiles with, Default
+#' @param col_palette Vector of colours to fill tiles with, Default
 #' `c("#4B3F72", "#CBB3BF", "#FFC857", "#119DA4", "#19647E")`.
 #' @param line_col Colour of lines between tiles, Default "white".
 #' @param line_size Thickness of lines between tiles. Default 1.
@@ -19,7 +19,7 @@
 #' @export
 
 mosaic <- function(n = 10,
-                   fill_cols = c("#4B3F72", "#CBB3BF", "#FFC857", "#119DA4", "#19647E"),
+                   col_palette = c("#4B3F72", "#CBB3BF", "#FFC857", "#119DA4", "#19647E"),
                    line_col = "white",
                    line_size = 1,
                    x_means = c(0, 10, 5),
@@ -35,7 +35,7 @@ mosaic <- function(n = 10,
         .f = ~ data.frame(x = rnorm(n, .x, xy_var), y = rnorm(n, .y, xy_var))
       )
       plot_data <- dplyr::bind_rows(df)
-      num_cols <- length(fill_cols)
+      num_cols <- length(col_palette)
       plot_data$fill_col <- factor(sample(1:num_cols,
         size = nrow(plot_data),
         replace = TRUE
@@ -57,7 +57,7 @@ mosaic <- function(n = 10,
       colour = line_col,
       size = line_size
     ) +
-    ggplot2::scale_fill_manual(values = fill_cols) +
+    ggplot2::scale_fill_manual(values = col_palette) +
     ggplot2::coord_cartesian(expand = FALSE) +
     theme_aRt(line_col)
   return(suppressWarnings(print(p)))